change data type to float

+++ main.py
@@ -56,4 +56,4 @@ if __name__ == '__main__':
     for i in range(epoch):
         for data in tqdm(train_loader):
             for step in gan_trainer.get_steps():
-                gan_trainer.train_step(step, data=data.to(working_device))
+                gan_trainer.train_step(step, data=data.to(working_device).float())
